requirements.txt
discord.py[voice]>=2.6.0
pynacl>=1.5.0
discord-ext-voice-recv
python-dotenv


# elevenlabs>=1.57.0
# openai>=1.63.0
# google-genai
# ffmpeg-python
# numpy
# scipy
# matplotlib
